Implement support for OsmosisLocal
This has been useful for basically running an osmosis node locally and
deploying a contract there and testing things out.

I tested this change against the internal QA tool and things seems to
be working fine.

The motivation of adding this is to for the on chain tests that we
were discussing.