Support backup chain nodes
We used to have only one pool with only one endpoint in Vec, and retry on it every where. Recent troubles showed that there might be some broken blocks in the certain node we are connecting, retry made no effect at all.

Hence this change makes full use of the framework, allowing more endpoints to be specified. And in `perform_query`, the primary pool (as `pool` used to be, contains only one endpoint) is used first. If it failed, more endpoints would be tried. Not working connections would be marked as broken. If all failed, errors from primary connection would be returned