Use a custom connection pool instead of bb8
Nothing wrong with bb8, but the fallback connection requirements mean we
have a heterogenous set of connections, and a standard connection pool
library just doesn't provide the facilities we want for managing it. We
had a bunch of wonky workarounds to deal with the fact that it was
difficult to mark specific endpoints as unhealthy during the connecting
phase. This should hopefully simplify things going forward.